SmoothieElement looks nice

+++ Runtime/Animation/SmoothieElementAnimationStyle.cs
@@ -14,8 +14,11 @@ namespace Smoothie
         [HideInInspector]
         public SmoothieElementAnimationManager ManagerRef;
 
+        [BoxGroup("Style Info", showLabel: false)]
+        [HorizontalGroup("Style Info/Name")]
         [SerializeField, OnValueChanged("UpdateAssetName")]
         [LabelText("Style Name")]
+        [LabelWidth(80)]
         private string _styleName = "Untitled";
 
         public string styleName
@@ -24,25 +27,42 @@ namespace Smoothie
             set => _styleName = value;
         }
 
+        [HorizontalGroup("Style Info/Name")]
+        [Button("Add Event", ButtonSizes.Small), GUIColor(0.4f, 0.8f, 0.4f)]
+        private void AddEventButton()
+        {
+            AddNewEvent();
+        }
+
+        [HorizontalGroup("Style Info/Name", Width = 120)]
+        [Button("Remove Style", ButtonSizes.Small), GUIColor(0.9f, 0.3f, 0.3f)]
+        private void RemoveStyleButton()
+        {
+            RemoveThisStyle();
+        }
+
         [ListDrawerSettings(
             Expanded = true,
-            DraggableItems = false,
-            HideAddButton = false,
-            HideRemoveButton = false)]
-        [LabelText("Event Overrides")]
+            DraggableItems = true,
+            HideAddButton = true,
+            ShowItemCount = true)]
+        [LabelText("Event Definitions")]
         public List<ElementAnimationEventDependent> eventDefinitions = new List<ElementAnimationEventDependent>();
 
         // Красная кнопка для удаления этого стиля
-        [Button("Remove This Style", ButtonHeight = 25)]
-        [GUIColor(1f, 0.25f, 0.25f)]
 #if UNITY_EDITOR
         private void RemoveThisStyle()
         {
             if (ManagerRef != null)
             {
-                ManagerRef.dependentStyles.Remove(this);
-                AssetDatabase.RemoveObjectFromAsset(this);
-                AssetDatabase.SaveAssets();
+                if (EditorUtility.DisplayDialog("Remove Style", 
+                    $"Are you sure you want to remove style '{_styleName}'?", 
+                    "Yes, Remove", "Cancel"))
+                {
+                    ManagerRef.dependentStyles.Remove(this);
+                    AssetDatabase.RemoveObjectFromAsset(this);
+                    AssetDatabase.SaveAssets();
+                }
             }
             else
             {
@@ -50,6 +70,21 @@ namespace Smoothie
             }
         }
 #endif
+
+        [Button("Add Event", ButtonSizes.Large), GUIColor(0.4f, 0.8f, 0.4f)]
+        private void AddNewEvent()
+        {
+            var newEvent = new ElementAnimationEventDependent();
+            newEvent.SetParentStyle(this);
+            
+            // Если есть доступные события, выбираем первое
+            if (ManagerRef != null && ManagerRef.possibleEvents.Count > 0)
+            {
+                newEvent.eventKey = ManagerRef.possibleEvents[0];
+            }
+            
+            eventDefinitions.Add(newEvent);
+        }
 
         public bool TryGetEventDefinition(string key, out ElementAnimationEventDependent def)
         {
@@ -81,6 +116,7 @@ namespace Smoothie
                 if (eventDefinitions[i] != null)
                 {
                     eventDefinitions[i].SetParentStyle(this);
+                    eventDefinitions[i].UpdateParentReferences();
                 }
             }
         }
